Creación del módulo de finanzas, es un estado de cuenta para la empre…

+++ frontend/src/app/componentes/dialogs/mantenimiento-dialog/mantenimiento-dialog.component.ts
@@ -3,9 +3,12 @@ import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 import { faEye, faTrashAlt } from '@fortawesome/free-solid-svg-icons';
 import { Subject } from 'rxjs';
 import { Mantenimiento } from 'src/app/interfaces/mantenimiento';
+import { Pago } from 'src/app/interfaces/pago';
 import { Propiedad } from 'src/app/interfaces/propiedad';
 import { AuthService } from 'src/app/servicios/auth.service';
 import { MantenimientosService } from 'src/app/servicios/mantenimientos.service';
+import { PagosService } from 'src/app/servicios/pagos.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-mantenimiento-dialog',
@@ -45,16 +48,22 @@ export class MantenimientoDialogComponent implements OnDestroy, OnInit {
     propiedadId: ''
   };
 
+  egresoActivo: Pago = {
+    _id: '',
+    fecha: '',
+    metodoPago: '', //Aquí se utilizará como conepto del egreso
+    monto: 0
+  }
+
   //Default methods
   constructor(@Inject(MAT_DIALOG_DATA) public data:{
     tituloVentana: string,
     propiedadActiva: Propiedad
-  }, private mttoService: MantenimientosService, public authService: AuthService) {
+  }, private mttoService: MantenimientosService, public authService: AuthService, public pagoService: PagosService) {
     this.propiedadActiva = data.propiedadActiva;
   }
 
   ngOnInit(): void {
-    this.GetMantenimientos();
     this.dtOptions = {
       pagingType: 'full_numbers',
       pageLength: 10,
@@ -63,6 +72,7 @@ export class MantenimientoDialogComponent implements OnDestroy, OnInit {
         url: '../../../assets/lang/datatable_lang.json'
       }
     };
+    this.GetMantenimientos();
   }
 
   ngOnDestroy(): void {
@@ -71,6 +81,47 @@ export class MantenimientoDialogComponent implements OnDestroy, OnInit {
   }
 
   //Custom Methods
+
+  ClearFields(){
+    this.egresoActivo._id = '';
+    this.egresoActivo.fecha = '';
+    this.egresoActivo.metodoPago = '';
+    this.egresoActivo.monto = 0;
+  }
+
+  MostrarError(mensaje: string){
+    $("#error_mtto").text(mensaje);
+    $("#error_mtto").hide();
+    $("#error_mtto").show("slow", function(){
+      setTimeout(function(){
+        $("#error_mtto").hide("slow");
+      }, 1800)
+    });
+  }
+
+  ValidarFormulario(){
+    const descripcion = $("#txtDescripcion").val();
+    const costo = $("#txtCosto").val();
+    const fecha = $("#dtFecha").val();
+    
+    if(descripcion == ""){
+      this.MostrarError("El campo descripción se encuentra vacío");
+      return false;
+    }
+
+    if(costo == "" || costo == "0"){
+      this.MostrarError("El campo costo se encuentra vacío");
+      return false;
+    }
+
+    if(fecha == ""){
+      this.MostrarError("El campo fecha se encuentra vacío");
+      return false;
+    }
+
+    return true;
+  }
+
   ClearFieldsMtto(){
     this.mantenimientoActivo._id = '';
     this.mantenimientoActivo.fecha = '';
@@ -79,6 +130,10 @@ export class MantenimientoDialogComponent implements OnDestroy, OnInit {
   }
   
   GetMantenimientos(){
+    if($.fn.dataTable.isDataTable('#dtMantenimientos')){
+      var table = $("#dtMantenimientos").DataTable();
+      table.destroy();
+    }
     this.mttoService.GetTodosByProp(this.propiedadActiva._id).subscribe(
       res => {
         this.mantenimientos = res;
@@ -95,7 +150,13 @@ export class MantenimientoDialogComponent implements OnDestroy, OnInit {
   AgregarMantenimiento(){
     this.mttoService.CrearMantenimiento(this.mantenimientoActivo).subscribe(
       res => {
-        alert("Mantenimiento agregado");
+        this.GenerarEgreso();
+        Swal.fire({
+          title: "Mantenimiento registrado", 
+          text: 'El mantenimiento se ha registrado exitosamente.',
+          icon: 'success',
+          confirmButtonText: 'Aceptar'
+        });
         this.ClearFieldsMtto();
         this.GetMantenimientos();
       },
@@ -114,6 +175,11 @@ export class MantenimientoDialogComponent implements OnDestroy, OnInit {
   }
 
   GuardarMantenimiento(){
+    //Validamos
+    if(!this.ValidarFormulario()){
+      return; //Detenemos ejecución
+    }
+
     this.mantenimientoActivo.propiedadId = this.propiedadActiva._id;
     if(this.mantenimientoActivo._id === ''){
       this.AgregarMantenimiento();
@@ -135,6 +201,11 @@ export class MantenimientoDialogComponent implements OnDestroy, OnInit {
   }
 
   EliminarMantenimiento(mttoId: string){
+    //Validamos
+    if(!this.ValidarFormulario()){
+      return; //Detenemos ejecución
+    }
+
     if(confirm("¿Realmente desea eliminar el mantenimiento?")){
       this.mttoService.DeleteMantenimiento(mttoId).subscribe(
         res => {
@@ -146,6 +217,20 @@ export class MantenimientoDialogComponent implements OnDestroy, OnInit {
         err => console.log(err)
       )
     }
+  }
+
+  GenerarEgreso(){
+    this.egresoActivo.fecha = this.mantenimientoActivo.fecha;
+    this.egresoActivo.metodoPago = "Mantenimiento: " + this.mantenimientoActivo.descripcion;
+    this.egresoActivo.monto = this.mantenimientoActivo.costo;
+
+    this.pagoService.CrearPago(this.egresoActivo, 0).subscribe(
+      res => {
+        this.ClearFields();
+        //this.matDialogRef.close(true);
+      },
+      err => console.log(err)
+    )
   }
 
 }
